requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
TA-Lib>=0.4.25
matplotlib>=3.7.0
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
